Refactor temp ramp, improve log messages, make IOPS uint64
Encode temperature color stops as celsius/color pairs so the ramp
handles uneven spacing. Add debug logging to refreshTemp, fix
duplicate defaultInterface comment, and make IOPS fields uint64
for consistency with other collector types.

+++ internal/sysinfo/sysinfo.go
@@ -15,7 +15,7 @@ type NetBandwidth struct {
 // DiskIO holds per-second disk throughput and IOPS.
 type DiskIO struct {
 	ReadBytesPerSec, WriteBytesPerSec uint64
-	ReadIOPS, WriteIOPS               uint32
+	ReadIOPS, WriteIOPS               uint64
 }
 
 // DietPiStatus indicates whether DietPi is installed and up to date.
